requirements.txt
asgiref==3.8.1
Bottleneck==1.4.2
crispy-bootstrap4==2024.10
Django==5.1.3
django-crispy-forms==2.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
et-xmlfile==1.1.0
mysqlclient==2.0.3
numexpr==2.10.1
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
pip==25.0
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.2
tzdata==2023.3
wheel==0.45.1
